refactor(tab-manager): decompose AiChat god component into focused modules
Split the monolithic 845-line chat.svelte.ts and 351-line AiChat.svelte
into 15 focused files with single responsibilities:

State layer (4 files):
- ai/providers.ts — pure provider/model config data
- ai/ui-message.ts — UIMessage boundary + drift detection
- state/conversation-handle.svelte.ts — per-conversation reactive handle with DI
- state/chat-state.svelte.ts — orchestrator (CRUD, active tracking, observers)

Component layer (11 files):
- AiChat.svelte — thin orchestrator (42 lines, down from 351)
- ConversationPicker, MessageList, MessageParts, ChatInput, ChatErrorBanner
- ToolCallPart, ToolResultPart, ThinkingPart (MessagePart rendering)
- ModelCombobox (now controlled), ProviderSelect (new)

Key changes:
- ConversationHandle uses ConversationHandleDeps for dependency injection
- updateConversation uses atomic TableHelper.update() instead of find-then-set
- ModelCombobox accepts value/models/onSelect props (no singleton import)
- All 8 MessagePart types now render (text, tool-call, tool-result, thinking,
  plus placeholders for image/audio/video/document)
- Tool calls show human-readable names, status badges, and collapsible details
