+++ apps/tab-manager/src/lib/components/ChatErrorBanner.svelte
@@ -1,0 +1,47 @@
+<script lang="ts">
+	import { Button } from '@epicenter/ui/button';
+	import RotateCcwIcon from '@lucide/svelte/icons/rotate-ccw';
+	import XIcon from '@lucide/svelte/icons/x';
+
+	let {
+		error,
+		dismissedError,
+		onRetry,
+		onDismiss,
+	}: {
+		error: Error | undefined;
+		dismissedError: string | null;
+		onRetry: () => void;
+		onDismiss: () => void;
+	} = $props();
+
+	const visible = $derived(error && error.message !== dismissedError);
+</script>
+
+{#if visible}
+	<div
+		role="alert"
+		class="flex items-center justify-between gap-2 border-t border-destructive/20 bg-destructive/10 px-3 py-2 text-xs text-destructive"
+	>
+		<span class="min-w-0 flex-1">{error?.message}</span>
+		<div class="flex shrink-0 items-center gap-1">
+			<Button
+				variant="ghost"
+				size="sm"
+				class="h-6 gap-1 px-2 text-xs text-destructive hover:text-destructive"
+				onclick={onRetry}
+			>
+				<RotateCcwIcon class="size-3" />
+				Retry
+			</Button>
+			<Button
+				variant="ghost"
+				size="icon"
+				class="size-6 text-destructive hover:text-destructive"
+				onclick={onDismiss}
+			>
+				<XIcon class="size-3" />
+			</Button>
+		</div>
+	</div>
+{/if}
