+++ apps/tab-manager/src/lib/components/ConversationPicker.svelte
@@ -1,0 +1,165 @@
+<script lang="ts">
+	import type { ConversationHandle } from '$lib/state/chat-state.svelte';
+	import type { ConversationId } from '$lib/workspace';
+	import { confirmationDialog } from '$lib/components/ConfirmationDialog.svelte';
+	import { useCombobox } from '@epicenter/ui/hooks';
+	import { Button } from '@epicenter/ui/button';
+	import { cn } from '@epicenter/ui/utils';
+	import * as Command from '@epicenter/ui/command';
+	import * as Popover from '@epicenter/ui/popover';
+	import ChevronDownIcon from '@lucide/svelte/icons/chevron-down';
+	import MessageSquarePlusIcon from '@lucide/svelte/icons/message-square-plus';
+	import CheckIcon from '@lucide/svelte/icons/check';
+	import LoaderCircleIcon from '@lucide/svelte/icons/loader-circle';
+	import TrashIcon from '@lucide/svelte/icons/trash';
+
+	let {
+		conversations,
+		activeId,
+		onSwitch,
+		onCreate,
+	}: {
+		conversations: ConversationHandle[];
+		activeId: ConversationId;
+		onSwitch: (id: ConversationId) => void;
+		onCreate: () => void;
+	} = $props();
+
+	const conversationPicker = useCombobox();
+	let conversationSearch = $state('');
+
+	const filteredConversations = $derived(
+		conversationSearch
+			? conversations.filter((c) =>
+					c.title.toLowerCase().includes(conversationSearch.toLowerCase()),
+				)
+			: conversations,
+	);
+
+	/** Whether there are any conversations to show in the dropdown. */
+	const hasConversations = $derived(conversations.length > 0);
+
+	/** Active conversation title for the header bar. */
+	const activeTitle = $derived(
+		conversations.find((c) => c.id === activeId)?.title ?? 'New Chat',
+	);
+
+	/** Format a timestamp as a short relative time string. */
+	function formatRelativeTime(ms: number): string {
+		const seconds = Math.floor((Date.now() - ms) / 1000);
+		if (seconds < 60) return 'now';
+		const minutes = Math.floor(seconds / 60);
+		if (minutes < 60) return `${minutes}m`;
+		const hours = Math.floor(minutes / 60);
+		if (hours < 24) return `${hours}h`;
+		const days = Math.floor(hours / 24);
+		if (days < 7) return `${days}d`;
+		return new Date(ms).toLocaleDateString(undefined, {
+			month: 'short',
+			day: 'numeric',
+		});
+	}
+</script>
+
+<div class="flex items-center gap-1 border-b px-2 py-1.5">
+	{#if hasConversations}
+		<Popover.Root bind:open={conversationPicker.open}>
+			<Popover.Trigger bind:ref={conversationPicker.triggerRef}>
+				{#snippet child({ props })}
+					<Button
+						{...props}
+						variant="ghost"
+						size="sm"
+						role="combobox"
+						aria-expanded={conversationPicker.open}
+						class="h-7 min-w-0 flex-1 justify-between gap-1 px-2 text-xs"
+					>
+						<span class="truncate">{activeTitle}</span>
+						<ChevronDownIcon class="size-3 shrink-0 opacity-50" />
+					</Button>
+				{/snippet}
+			</Popover.Trigger>
+			<Popover.Content class="w-[280px] p-0" align="start">
+				<Command.Root shouldFilter={false}>
+					<Command.Input
+						placeholder="Search conversations…"
+						class="h-9 text-sm"
+						bind:value={conversationSearch}
+					/>
+					<Command.List class="max-h-[300px]">
+						<Command.Empty>No conversations found.</Command.Empty>
+						{#each filteredConversations as conv (conv.id)}
+							<Command.Item
+								value={conv.id}
+								class="group flex-col items-start gap-0.5"
+								onSelect={() => {
+									onSwitch(conv.id);
+									conversationSearch = '';
+									conversationPicker.closeAndFocusTrigger();
+								}}
+							>
+								<span
+									class="flex w-full items-center justify-between gap-1.5 text-xs"
+								>
+									<span class="flex min-w-0 items-center gap-1.5">
+										<CheckIcon
+											class={cn('mr-0.5 size-3 shrink-0', {
+												'text-transparent': conv.id !== activeId,
+											})}
+										/>
+										<span class="min-w-0 truncate font-medium"
+											>{conv.title}</span
+										>
+										{#if conv.isLoading}
+											<LoaderCircleIcon
+												class="size-3 shrink-0 animate-spin text-muted-foreground"
+											/>
+										{/if}
+									</span>
+									<span class="flex shrink-0 items-center gap-1">
+										<span class="text-[10px] text-muted-foreground"
+											>{formatRelativeTime(conv.updatedAt)}</span
+										>
+										<button
+											class="shrink-0 rounded p-0.5 opacity-0 transition-opacity hover:bg-destructive/10 hover:text-destructive group-hover:opacity-100"
+											onclick={(e) => {
+												e.stopPropagation();
+												e.preventDefault();
+												confirmationDialog.open({
+													title: 'Delete conversation',
+													description: `Delete "${conv.title}"? This will remove all messages in this conversation.`,
+													confirm: { text: 'Delete', variant: 'destructive' },
+													onConfirm: () => conv.delete(),
+												});
+											}}
+										>
+											<TrashIcon class="size-3" />
+										</button>
+									</span>
+								</span>
+								{@const preview = conv.lastMessagePreview}
+								{#if preview}
+									<span
+										class="w-full truncate pl-5 text-[10px] text-muted-foreground"
+										>{preview}</span
+									>
+								{/if}
+							</Command.Item>
+						{/each}
+					</Command.List>
+				</Command.Root>
+			</Popover.Content>
+		</Popover.Root>
+	{:else}
+		<span class="flex-1 px-2 text-xs text-muted-foreground">No chats yet</span>
+	{/if}
+
+	<Button
+		variant="ghost"
+		size="icon"
+		class="size-7 shrink-0"
+		onclick={() => onCreate()}
+	>
+		<MessageSquarePlusIcon class="size-3.5" />
+	</Button>
+</div>
